requirements.txt
matplotlib==3.9.2
numpy==2.1.3
PyYAML==6.0.2
tqdm==4.66.4
